+++ django_micro.py
@@ -1,30 +1,46 @@
-from __future__ import absolute_import
-
+from importlib import import_module
 import inspect
-import os
 import sys
+import os
 
 import django
 from django.conf import settings
-from django.conf.urls import url
 from django.core import management
-from django.core.exceptions import ImproperlyConfigured
 from django.template import Library
+from django.urls import path, re_path
+from django.core.management.base import BaseCommand
+from django.core.exceptions import ImproperlyConfigured
 
-__all__ = ['command', 'configure', 'run', 'route', 'template']
+__all__ = ['command', 'configure', 'run', 'route', 'template', 'get_app_label']
 
-
-# -------------------
-# Application module
-# -------------------
+register = template = Library()
+urlpatterns = []
 
 _app_module = None
 _app_label = None
+_commands = {}
 
 
-def _setup_module():
-    app_module = sys.modules[inspect.stack()[2][0].f_locals['__name__']]
-    app_label = os.path.basename(os.path.dirname(os.path.abspath(app_module.__file__)))
+def _create_app(stack):
+    # extract directory, filename and app label from parent module
+    app_module = sys.modules[stack[1][0].f_locals['__name__']]
+    app_dirname = os.path.dirname(os.path.abspath(app_module.__file__))
+    app_file_name = os.path.basename(app_module.__file__).split('.')[0]
+    app_label = os.path.basename(app_dirname)
+
+    app_module_name = '{}.{}'.format(app_label, app_file_name)
+
+    # use parent directory of application as import root
+    sys.path[0] = os.path.dirname(app_dirname)
+
+    # allow relative import from app.py
+    app_module.__package__ = app_label
+    import_module(app_label)
+
+    if app_module.__name__ != app_module_name:
+        # allow recursive import app.py
+        sys.modules[app_module_name] = app_module
+
     globals().update(_app_module=app_module, _app_label=app_label)
 
 
@@ -36,49 +52,39 @@ def get_app_label():
     return _app_label
 
 
-# -------------------
-# Views and routes
-# -------------------
+def route(pattern, view_func=None, regex=False, *args, **kwargs):
+    path_func = re_path if regex else path
 
-urlpatterns = []
-
-
-def route(pattern, view_func=None, *args, **kwargs):
     def decorator(view_func):
         if hasattr(view_func, 'as_view'):
             view_func = view_func.as_view()
-        urlpatterns.append(url(pattern, view_func, *args, **kwargs))
+
+        urlpatterns.append(path_func(pattern, view_func, *args, **kwargs))
         return view_func
 
     # allow use decorator directly
-    # route(r'^$', show_index)
+    # route('blog/', show_index)
     if view_func:
         return decorator(view_func)
 
     return decorator
 
 
-# -------------------
-# Template tags
-# -------------------
+def configure(config_dict={}, django_admin=False):
+    _create_app(inspect.stack())  # load application from parent module
 
-register = template = Library()
+    if 'BASE_DIR' in config_dict:
+        config_dict.setdefault('TEMPLATE_DIRS', [os.path.join(config_dict['BASE_DIR'], 'templates')])
 
+    if django_admin:
+        _configure_admin(config_dict)
 
-# --------------------
-# Configuration
-# --------------------
-
-def configure(config_dict={}, enable_admin=False):
-    _setup_module()  # find in the stack module that calls this function
-    config_dict.setdefault('TEMPLATE_DIRS', ['templates'])
-
-    kwargs = {
+    django_config = {
         'INSTALLED_APPS': [_app_label] + config_dict.pop('INSTALLED_APPS', []),
         'ROOT_URLCONF': __name__,
         'TEMPLATES': [{
             'BACKEND': 'django.template.backends.django.DjangoTemplates',
-            'DIRS': config_dict.pop('TEMPLATE_DIRS'),
+            'DIRS': config_dict.pop('TEMPLATE_DIRS', []),
             'APP_DIRS': True,
             'OPTIONS': {
                 'context_processors': config_dict.pop('CONTEXT_PROCESSORS', []),
@@ -87,43 +93,40 @@ def configure(config_dict={}, enable_admin=False):
         }],
     }
 
-    if enable_admin:
-        admin_apps = [
+    django_config.update({key: val for key, val in config_dict.items() if key.isupper()})
+    settings.configure(**django_config)
+    django.setup()
+
+
+def _configure_admin(config_dict):
+    admin_deps = {
+        'INSTALLED_APPS': [
             'django.contrib.admin',
             'django.contrib.auth',
             'django.contrib.contenttypes',
             'django.contrib.sessions',
             'django.contrib.staticfiles',
-        ]
-        middlewares = [
+        ],
+        'MIDDLEWARE': [
             'django.contrib.sessions.middleware.SessionMiddleware',
             'django.contrib.auth.middleware.AuthenticationMiddleware',
             'django.middleware.security.SecurityMiddleware',
             'django.middleware.common.CommonMiddleware',
-        ]
-        kwargs['INSTALLED_APPS'].extend(admin_apps)
-        if 'MIDDLEWARE' in kwargs.keys():
-            kwargs['MIDDLEWARE'].extend(middlewares)
-        else:
-            kwargs['MIDDLEWARE'] = middlewares
-        kwargs['TEMPLATES'][0]['OPTIONS']['context_processors'].append(
-            'django.contrib.auth.context_processors.auth'
-        )
-        kwargs.setdefault('STATIC_URL', '/static/')
+            'django.contrib.messages.middleware.MessageMiddleware',
+        ],
+        'CONTEXT_PROCESSORS': [
+            'django.contrib.auth.context_processors.auth',
+        ],
+    }
 
-    kwargs.update({key: val for key, val in config_dict.items() if key.isupper()})
-    settings.configure(**kwargs)
-    django.setup()
+    for key, value in admin_deps.items():
+        prev_value = config_dict.get(key, [])
+        config_dict[key] = value + prev_value
+
+    config_dict.setdefault('STATIC_URL', '/static/')
 
 
-# --------------------
-# Management commands
-# --------------------
-
-_commands = {}
-
-
-def patch_get_commands():
+def _patch_get_commands():
     django_get_commands = management.get_commands
 
     def patched_get_commands():
@@ -135,16 +138,32 @@ def patch_get_commands():
     management.get_commands = patched_get_commands
 
 
-def command(name, command_cls=None):
+def command(name=None, command_cls=None):
     if not getattr(management.get_commands, 'patched', False):
-        patch_get_commands()
+        _patch_get_commands()
+
+    if inspect.isfunction(name):
+        # Shift arguments if decroator called without brackets
+        command_cls = name
+        name = None
 
     def decorator(command_cls):
-        command_instance = command_cls()
-        # very dirty hack for extracting app name
-        # from command (via https://goo.gl/1c1Irj)
+        command_name = name
+
+        if inspect.isclass(command_cls):
+            command_instance = command_cls()
+        else:
+            # transform function-based command to class
+            command_name = name or command_cls.__name__
+            command_instance = type('Command', (BaseCommand,), {'handle': command_cls})()
+
+        if not command_name:
+            raise DjangoMicroException("Class-based commands requires name argument.")
+
+        # Hack for extracting app name from command (https://goo.gl/1c1Irj)
         command_instance.rpartition = lambda x: [_app_label]
-        _commands[name] = command_instance
+
+        _commands[command_name] = command_instance
         return command_cls
 
     # allow use decorator directly
@@ -155,9 +174,9 @@ def command(name, command_cls=None):
     return decorator
 
 
-# --------------------
-# Bootstrap
-# --------------------
+class DjangoMicroException(Exception):
+    pass
+
 
 def run():
     if not settings.configured: